Unite _led.py and led.py
